feat: Move prefersReducedMotion and getArticleTitleFromMain functions to dom-utils.js; refactor renderPostMetaCard to templates.js for better organization
